Extract shrink to separate file
